Moved all of the logging logic inside engine.logs
